Change cli.CommandLineInterface._arrow_date() to static method.
This resolves a subtle bug that was surfaced by attrs-17.1.0
(see http://www.attrs.org/en/latest/changelog.html#id83).

The consequence of this change for production deployments is that they
can use any attrs version >=16.3.0.

In development, this change allows for the use of the black code
formatting tool that requires attrs>=17.4.0 presently and
attrs>=18.1.0 in its next release.